+++ src/index.ts
@@ -6,3 +6,4 @@ export * from "./repair.js";
 export * from "./storyboard.js";
 export * from "./development.js";
 export * from "./version.js";
+export * from "./evaluation.js";